pyproject.toml
[tool.poetry]
name = "scenic"
version = "2.0.0"
description = "The Scenic scenario description language."
authors = [
	"Daniel Fremont <dfremont@ucsc.edu>",
	"Edward Kim",
	"Tommaso Dreossi",
	"Shromona Ghosh",
	"Xiangyu Yue",
	"Alberto L. Sangiovanni-Vincentelli",
	"Sanjit A. Seshia"
]
maintainers = ["Daniel Fremont <dfremont@ucsc.edu>"]
readme = "README.md"
license = "BSD-3-Clause"
repository = "https://github.com/BerkeleyLearnVerify/Scenic"
documentation = "https://scenic-lang.readthedocs.io"
classifiers = [
	'Operating System :: OS Independent',
	'Development Status :: 5 - Production/Stable',
]

[tool.poetry.dependencies]
python = "^3.8"
dotmap = "^1.3.13"
shapely = "^1.7.0"
mapbox_earcut = "^0.12.10"
matplotlib = "^3.2.1"
scipy = "^1.4.1"
antlr4-python3-runtime = "~4.8"
opencv-python = "~4.2.0.34"
numpy = "^1.18.2"
pillow = "^7.1.1"
pynverse = "^0.1.4"
pygame = "^2.0.0.dev6"
attrs = "^19.3.0"
wrapt = "^1.12.1"

pyproj = {version = "^2.6.0", optional = true}

pytest-randomly = {version = "^3.2.1", optional = true}
sphinx = {version = "^3.1.0", optional = true}
tox = {version = "^3.14.0", optional = true}
sphinx_rtd_theme = {version = "^0.4.3", optional = true}
astor = {version = "^0.8.1", optional = true}
pytest = "^6.2.1"
jupyter = "^1.0.0"
notebook = "^6.1.6"
statistics = "^1.0.3"
nuscenes-devkit = "^1.1.9"

[tool.poetry.extras]
guideways = ["pyproj"]
dev = [
	"pytest-randomly",
	"pytest",
	"sphinx",
	"sphinx_rtd_theme",
	"tox",
	"pygments",
	"astor",
	"pyproj",	# for running guideways modules
]

[tool.poetry.scripts]
scenic = 'scenic.__main__:dummy'

[tool.poetry.plugins."pygments.lexers"]
scenic = "scenic.syntax.pygment:ScenicLexer"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
